requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.4
python-multipart==0.0.20
boto3==1.35.90
python-dotenv==1.0.1
pillow==11.0.0